requirements.txt
streamlit
numpy
Pillow
cairosvg